academic-search-main/requirements-fulltext.txt
psycopg2-binary
requests
pypdf
